Improve API request logging middleware prefixes.
Specify target so that logs have a prefix for the module itself, not the
file the middleware was defined in.
