+++ src/api/mod.rs
@@ -2,4 +2,5 @@
 // Licensed under the Business Source License 1.1. See LICENSE file for details.
 
 pub mod admin;
+pub(crate) mod middleware;
 pub mod primary;
